# docs/swagger.yaml
basePath: /v1
definitions:
  backend-go_internal_core_models.Feature:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_deleted:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  backend-go_internal_core_models.Role:
    properties:
      RoleFeature:
        items:
          $ref: '#/definitions/backend-go_internal_core_models.RoleFeature'
        type: array
      active:
        type: boolean
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_deleted:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
  backend-go_internal_core_models.RoleFeature:
    properties:
      activate:
        type: boolean
      create:
        type: boolean
      delete:
        type: boolean
      id_feature:
        type: string
      id_role:
        type: string
      view:
        type: boolean
    type: object
...
  internal_app_role.CreateRoleDTO:
    properties:
      description:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/backend-go_internal_core_models.RoleFeature'
        type: array
    required:
    - description
    - name
    type: object
  internal_app_role.RoleListAllResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/backend-go_internal_core_models.Role'
        type: array
      total:
        type: integer
    type: object
  internal_app_role.RoleListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/backend-go_internal_core_models.Role'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  internal_app_role.UpdateStatusRequest:
    properties:
      active:
        type: boolean
    type: object
...
host: localhost:8888
info:
  contact:
    email: suporte@swagger.io
    name: Suporte API
    url: http://www.swagger.io/support
  description: API modular em Go com Gin e Swagger.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Backend Go API
  version: "1.0"
paths:
...
  /role:
    get:
      description: Retorna papéis com paginação e filtros.
      parameters:
      - description: 'Número da página (padrão: 1)'
        in: query
        name: page
        type: integer
      - description: 'Itens por página (padrão: 25)'
        in: query
        name: size
        type: integer
      - description: Termo de busca
        in: query
        name: searchWord
        type: string
      - description: 'Campos para busca (ex: name)'
        in: query
        name: searchFields
        type: string
      - description: Campo para ordenação
        in: query
        name: orderBy
        type: string
      - description: Direção da ordenação (asc/desc)
        in: query
        name: orderDirection
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_app_role.RoleListResponse'
      security:
      - Bearer: []
      summary: Listar Papéis (Roles)
      tags:
      - Role
    post:
      consumes:
      - application/json
      description: Cria um novo papel com permissões específicas.
      parameters:
      - description: Dados do papel
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_app_role.CreateRoleDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/backend-go_internal_core_models.Role'
        "400":
          description: Dados inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Criar Papel (Role)
      tags:
      - Role
  /role/{id}:
    delete:
      description: Remove um papel baseado no ID.
      parameters:
      - description: ID do Papel
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: papel excluído com sucesso'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Excluir Papel (Role)
      tags:
      - Role
    get:
      description: Retorna os detalhes de um papel e suas permissões.
      parameters:
      - description: ID do Papel
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend-go_internal_core_models.Role'
        "404":
          description: Papel não encontrado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Obter Papel (Role)
      tags:
      - Role
    put:
      consumes:
      - application/json
      description: Atualiza os dados e permissões de um papel baseado no ID.
      parameters:
      - description: ID do Papel
        in: path
        name: id
        required: true
        type: string
      - description: Dados do papel
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_app_role.CreateRoleDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend-go_internal_core_models.Role'
        "400":
          description: Dados inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Atualizar Papel (Role)
      tags:
      - Role
  /role/{id}/status:
    patch:
      consumes:
      - application/json
      description: Define se um papel está ativo ou inativo.
      parameters:
      - description: ID do Papel
        in: path
        name: id
        required: true
        type: string
      - description: Novo status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_app_role.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: status atualizado com sucesso'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Alterar Status do Papel (Role)
      tags:
      - Role
  /role/all:
    get:
      description: Retorna todos os papéis (usado para selects/lookups).
      parameters:
      - description: Termo de busca
        in: query
        name: searchWord
        type: string
      - description: Campos para busca
        in: query
        name: searchFields
        type: string
      - description: Campo para ordenação
        in: query
        name: orderBy
        type: string
      - description: Direção da ordenação
        in: query
        name: orderDirection
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_app_role.RoleListAllResponse'
      security:
      - Bearer: []
      summary: Listar Todos os Papéis (Roles)
      tags:
      - Role
  /role/features:
    get:
      description: Retorna a lista de todas as funcionalidades disponíveis para atribuição
        de permissões.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/backend-go_internal_core_models.Feature'
            type: array
        "500":
          description: Erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Listar Funcionalidades
      tags:
      - Role
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"